Prediction/requirements.txt
numpy==1.26.4
fastapi==0.115.4
pydantic==2.9.2
uvicorn==0.32.0
joblib==1.4.2
scikit-learn==1.5.2
pandas==2.2.3
scipy==1.13.1
contourpy==1.3.2
dm-tree==0.1.8
ml-dtypes==0.5.0
matplotlib==3.9.2
protobuf==5.28.2
requests==2.32.3
werkzeug==3.0.4
blinker==1.8.2
webencodings==0.5.1
markupsafe==3.0.2
packaging==24.1
pygments==2.18.0
setuptools==75.1.0
cython==3.0.11
PyYAML
